Add files via upload
the data base is linked up for logins and account creation but the program lost some functionality as a result. I will review this tomorrow.  
to run use the usual install and add npm install firebase, and npm install vuex@next 
I may have forgot an install on the instructions so don't hesitate to ask. 

note: this is only the source code.